% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beta_to_M.R
\name{Beta_to_M}
\alias{Beta_to_M}
\title{Beta_to_M}
\usage{
Beta_to_M(betas)
}
\arguments{
\item{betas}{A numeric vector containing the Beta values}
}
\value{
M values
}
\description{
Beta values represent the level of methylation on a scale of 0 and 100, where 0 means 0% methylation and 100 means 100% methylation. Despite the interpretability of Beta values, Du et al. 2010 argue that Beta values are more heteroscedastic than M-values, making them less reliable for downstream statistical analysis. This function converts Beta effect sizes into mean M effect sizes for methylation analysis.
}
\examples{
Beta <- c(0.74125143,0.93553086,0.76974692,0.83446278,0.71556639,0.05195812,0.05815754,0.33628677,0.73010305,0.40927036)
Mvals <- Beta_to_M(Beta)
}
\author{
Andrew Y.F. Li Yim
}
\keyword{Beta,}
\keyword{M-values}
