% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpg_dotbox_plot.R
\name{cpg_dotbox_plot}
\alias{cpg_dotbox_plot}
\title{cpg_dotbox_plot}
\usage{
cpg_dotbox_plot(cpg_num, betas, factor_interest, title, gg.plot = T,
  enlarged = "auto")
}
\arguments{
\item{cpg_num}{CpG number (ID)}

\item{betas}{Matrix of Beta with the CpG numbers as rownames}

\item{factor_interest}{A vector based on which you want to stratify your plot}

\item{gg.plot}{Do you want to plot using ggplot2 (TRUE) or base R plotting (FALSE). Defaults to TRUE.}

\item{enlarged}{Do you want an enlarged plot on the right side (TRUE or FALSE). Defaults to "auto" whereby the standard deviation will be calculated of the sample. If the sd is smaller than 0.015 the enlarged plot will be generated for visibility purposes.}
}
\value{
A boxplot overlayed by a stripplot of a specific CpG
}
\description{
Produces a boxplot overlayed by a stripplot of a specific CpG given a matrix of beta-values. Ensure that the Beta-matrix has the CpG number as the row names
}
\examples{
Beta <- matrix(c(0.15,0.2,0.17,0.76,0.8,0.65,0.22,0.23,0.24,0.5,0.51,0.52), nrow = 2,ncol = 6, byrow = T)
rownames(Beta) <- c("cg123456", "cg789010")
Pheno <- c(rep("pheno1", 3), rep("pheno2", 3))
cpg_dotbox_plot("cg123456", Beta, Pheno)
}
\author{
Andrew Y.F. Li Yim
}
\keyword{boxplot,}
\keyword{cpg,}
\keyword{jitter,}
\keyword{methylation}

