% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpg_strip_plot.R
\name{cpg_strip_plot}
\alias{cpg_strip_plot}
\title{cpg_strip_plot}
\usage{
cpg_strip_plot(cpg_num, betas, factor_interest, title, enlarged = "auto",
  type = c("boxplot", "CI", "SE", "mean"), colors = NULL,
  legend = TRUE, y_lab = NULL)
}
\arguments{
\item{cpg_num}{CpG number (ID)}

\item{betas}{Matrix of Beta with the CpG numbers as rownames}

\item{factor_interest}{A vector based on which you want to stratify your plot}

\item{enlarged}{Do you want an enlarged plot on the right side (TRUE or FALSE). Defaults to "auto" whereby the standard deviation will be calculated of the sample. If the sd is smaller than 0.015 the enlarged plot will be generated for visibility purposes.}

\item{type}{Choose the type of plot you want: "boxplot", "SE", "CI" or simply "mean". Defaults to "boxplot".}

\item{colors}{A vector containing the the colors to be utilized for the plot with the contrast of interest as names (default: uses ggplot's default coloring scheme).}

\item{legend}{Boolean depicting whether a legend should be included (default: TRUE)}

\item{y_lab}{A specific name for the y labels (default: "Beta")}
}
\value{
A boxplot overlayed by a stripplot of a specific CpG
}
\description{
Produces a boxplot overlayed by a stripplot of a specific CpG given a matrix of beta-values. Ensure that the Beta-matrix has the CpG number as the row names
}
\examples{
Beta <- matrix(c(0.15,0.2,0.17,0.76,0.8,0.65,0.22,0.23,0.24,0.5,0.51,0.52), nrow = 2,ncol = 6, byrow = T)
rownames(Beta) <- c("cg123456", "cg789010")
Pheno <- c(rep("pheno1", 3), rep("pheno2", 3))
cpg_strip_plot("cg123456", Beta, Pheno, type = "SE")
}
\author{
Andrew Y.F. Li Yim
}
\keyword{boxplot,}
\keyword{cpg,}
\keyword{methylation}
\keyword{stripplot,}
