% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpgs_within_region.R
\name{cpgs_within_region}
\alias{cpgs_within_region}
\title{cpgs_within_region}
\usage{
cpgs_within_region(chr, annotation.gr, start_region, end_region)
}
\arguments{
\item{chr}{Chromosome (example "chr6")}

\item{annotation.gr}{GenomicRanges object containing the annotation/features of the GMset (obtained through minfi::getAnnotation())}

\item{start_region}{The start of the region you want to plot}

\item{end_region}{The end of the region you want to plot. If the region is only 1 bp, simply enter the same value as the start_dmr}
}
\value{
A region of CpG(s) within the range of interest
}
\description{
Produces a dotplot of a specific (region of) CpG(s) given a matrix of beta-values superposed onto the RefSeq genomic location.
}
\examples{
#Load data
require(minfiData)
baseDir <- system.file("extdata", package = "minfiData")
targets <- read.450k.sheet(base = baseDir)
RGset <- read.450k.exp(targets = targets, recursive = T)
Mset <- preprocessIllumina(rgSet = RGset, bg.correct = T, normalize = "controls", reference = 2)
Rset <- ratioConvert(Mset)
GMset <- mapToGenome(Rset)

Beta <- getBeta(GMset)
annotation.gr <- makeGRangesFromDataFrame(getAnnotation(GMset), keep.extra.columns = T, start.field = "pos", end.field = "pos")

#Find DMRs
design <- model.matrix(~pData(Mset)$Sample_Group)
bumps <- bumphunter(object = GMset, design = design, coef = 2, cutoff = 0.2, B = 0, type = "Beta")
dmr <- bumps$table[1,]

#Find DMPs within DMRs
cpg_within_region(chr = "chr6", annotation.gr = annotation.gr, start_dmr = dmr$start, end_dmr = dmr$start)

}
\author{
Andrew Y.F. Li Yim
}
\keyword{cpg,}
\keyword{dmr,}
\keyword{methylation}

