% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr_genome_plot.R
\name{dmr_genome_plot}
\alias{dmr_genome_plot}
\title{dmr_genome_plot}
\usage{
dmr_genome_plot(name, chr, betas, annotation.gr, factor_interest, start_dmr,
  end_dmr, flanks, SNP = F, Reg = F, dmps = F, genome_version = "hg19",
  diff_symbol = T, dotsize = 0.7, plotgrid = T, confint = T,
  highlight = T, col_DMR = NULL)
}
\arguments{
\item{name}{Name of the DMR/DMP}

\item{chr}{Chromosome (example "chr6")}

\item{betas}{Matrix of Beta with the CpG numbers as rownames}

\item{annotation.gr}{GenomicRanges object containing the annotation/features of the GMset (obtained through minfi::getAnnotation())}

\item{factor_interest}{A vector based on which you want to stratify your plot}

\item{start_dmr}{The start of the region you want to plot}

\item{end_dmr}{The end of the region you want to plot. If the region is only 1 bp, simply enter the same value as the start_dmr}

\item{flanks}{How large should the flanking region be? Defaults to the same size as the DMR.}

\item{SNP}{Do you want SNP information sourced from UCSC? Defaults to FALSE, because calling UCSC is slow.}

\item{Reg}{Do you want Regulatory information sourced from UCSC? Defaults to FALSE, because calling UCSC is slow.}

\item{dmps}{Do you want to indicate them individually? Defaults to FALSE}

\item{genome_version}{Which genome build to use? Defaults to "hg19"}

\item{diff_symbol}{Do you want the different groups to be represented by different symbols (TRUE)? Defaults to FALSE}

\item{dotsize}{The pixel size for plotting symbols. Defaults to 0.7 (standard of Gviz)}

\item{plotgrid}{Do you want to plot a grid (TRUE)? Defaults to TRUE}

\item{confint}{Do you want to plot the 95\% confidence intervals (TRUE)? Defaults to TRUE}

\item{highlight}{Do you want to highlight the region of interest (TRUE)? Defaults to TRUE}

\item{col_DMR}{A color vector used for plotting the DMRs. Must be the same size as the levels of factor_interest. Defaults to NULL (which is automatic color assignment)}
}
\value{
A dotplot of a specific (region of) CpG(s) of beta-values superposed onto the RefSeq genome. Additionally, it returns a genomic ranges object with the CpGs within the chosen region.
}
\description{
Produces a dotplot of a specific (region of) CpG(s) given a matrix of beta-values superposed onto the RefSeq genomic location.
}
\examples{
#Load data
require(minfiData)
baseDir <- system.file("extdata", package = "minfiData")
targets <- read.metharray.sheet(base = baseDir)
RGset <- read.metharray.exp(targets = targets, recursive = T)
Mset <- preprocessIllumina(rgSet = RGset, bg.correct = T, normalize = "controls", reference = 2)
Rset <- ratioConvert(Mset)
GMset <- mapToGenome(Rset)

Beta <- getBeta(GMset)
annotation.gr <- makeGRangesFromDataFrame(getAnnotation(GMset), keep.extra.columns = T, start.field = "pos", end.field = "pos")

#Find DMRs
design <- model.matrix(~pData(Mset)$Sample_Group)
bumps <- bumphunter(object = GMset, design = design, coef = 2, cutoff = 0.2, B = 0, type = "Beta")
dmr <- bumps$table[1,]

#Plot DMP
dmr_genomeplot(name = "DMR_1", chr = "chr6", betas = Beta, factor_interest = pData(Mset)$Sample_Group, annotation.gr = annotation.gr, start_dmr = dmr$start, end_dmr = dmr$start)

#Plot DMR (without SNPs)
dmr_genomeplot(name = "DMR_1", chr = "chr6", betas = Beta, factor_interest = pData(Mset)$Sample_Group, annotation.gr = annotation.gr, start_dmr = dmr$start, end_dmr = dmr$end)

#Plot DMR (with SNPs)
dmr_genome_plot(name = "DMR_1", chr = "chr6", betas = Beta, factor_interest = pData(Mset)$Sample_Group, annotation.gr = annotation.gr, start_dmr = dmr$start, end_dmr = dmr$end, SNP = T) 
}
\seealso{
\pkg{\link{minfi}}

\pkg{\link{minfiData}}
}
\author{
Andrew Y.F. Li Yim
}
\keyword{cpg,}
\keyword{dmr,}
\keyword{methylation}
