% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_cpg_finder.R
\name{gene_cpg_finder}
\alias{gene_cpg_finder}
\title{gene_cpg_finder}
\usage{
gene_cpg_finder(gene_symbol, annotation.gr)
}
\arguments{
\item{gene_symbol}{A (list of) UCSC gene symbol(s)}

\item{annotation.gr}{GenomicRanges object containing the annotation/features of the GMset (obtained through minfi::getAnnotation())}
}
\value{
A (list of) GenomicRanges object(s) containing the CpG annotations associated to the gene of interest
}
\description{
Find the probes associated to a given gene symbol (UCSC) using Regex
}
\examples{
#Load data
require(minfiData)
baseDir <- system.file("extdata", package = "minfiData")
targets <- read.450k.sheet(base = baseDir)
RGset <- read.450k.exp(targets = targets, recursive = T)
Mset <- preprocessIllumina(rgSet = RGset, bg.correct = T, normalize = "controls", reference = 2)
Rset <- ratioConvert(Mset)
GMset <- mapToGenome(Rset)

annotation.gr <- makeGRangesFromDataFrame(getAnnotation(GMset), keep.extra.columns = T, start.field = "pos", end.field = "pos")

gene_cpg_finder(gene_symbol = "TNF", annotation.gr = annotation.gr)
}
\author{
Andrew Y.F. Li Yim
}
\keyword{cpg,}
\keyword{methylation}
