% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmp.R
\name{lmp}
\alias{lmp}
\title{lmp}
\usage{
lmp(modelobject)
}
\arguments{
\item{modelobject}{lm output}
}
\value{
A p-value for a linear regression as obtained through lm()
}
\description{
Extracts the p-value from lm(). Source: Stephen Turner (http://www.gettinggeneticsdone.com/2011/01/rstats-function-for-extracting-f-test-p.html)
In short, lmp extracts the F-statistic from the lm output and calculates the p-value based on the F-distribution.
}
\examples{
set.seed(42)
n=20
d=data.frame(x1=rbinom(n,2,.5), x2=rbinom(n,2,.5))
d=transform(d, y=x1+x2+rnorm(n))

#fit the linear model
fit=lm(y ~ x1 + x2, data=d)

summary(fit) #shows that the F-test is 0.006641
names(summary(fit)) #can't access that p-value using this!
names(fit) # this doesn't work either
lmp(fit) # uses the above function to capture the F-test p-value.
#Source: http://www.gettinggeneticsdone.com/2011/01/rstats-function-for-extracting-f-test-p.html
}
\author{
Stephen Turner
}
\keyword{fstatistic}
\keyword{lm,}
\keyword{p-value,}

