% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_to_Beta.R
\name{M_to_Beta}
\alias{M_to_Beta}
\title{M_to_Beta}
\usage{
M_to_Beta(mvals)
}
\arguments{
\item{mvals}{A numeric vector containing the M-values}
}
\value{
Beta values
}
\description{
Beta values represent the level of methylation on a scale of 0 and 100, where 0 means 0% methylation and 100 means 100% methylation. Despite the interpretability of Beta values, Du et al. 2010 argue that Beta values are more heteroscedastic than M-values, making them less reliable for downstream statistical analysis. This function converts mean M values into their respective Beta values for methylation analysis.
}
\examples{
Mvals <- c(1.5184121,3.8591045,1.7411637,2.3336920,1.3309937,-4.1895298,-4.0174476,-0.9808680,1.4356915,-0.5294438)
Betas <- MtoBeta(Mvals)
}
\author{
Andrew Y.F. Li Yim
}
\keyword{Beta,}
\keyword{M-values}
