% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methdata_to_granges.R
\name{methdata_to_granges}
\alias{methdata_to_granges}
\title{methdata_to_granges}
\usage{
methdata_to_granges(
  methdata,
  annotation.gr,
  chr,
  start_region,
  end_region,
  flanks
)
}
\arguments{
\item{methdata}{Methylation data (betas or M-values)}

\item{annotation.gr}{Annotation data in GenomicRanges format}

\item{chr}{The chromosome as character (ex. "chr6")}

\item{start_region}{The start of the region as numeric}

\item{end_region}{The end of the region as numeric}

\item{flanks}{The size of the around the region of interest. Defaults to 0}
}
\value{
A region of the annotation.gr and betas given the coordinates in GenomicRanges format
}
\description{
Produces a dotplot of a specific (region of) CpG(s) given a matrix of beta-values superposed onto the RefSeq genomic location.
}
\examples{
#Load data
require(minfiData)
baseDir <- system.file("extdata", package = "minfiData")
targets <- read.metharray.sheet(base = baseDir)
RGset <- read.metharray.exp(targets = targets, recursive = T)
Mset <- preprocessIllumina(rgSet = RGset, bg.correct = T, normalize = "controls", reference = 2)
Rset <- ratioConvert(Mset)
GMset <- mapToGenome(Rset)

Beta <- getBeta(GMset)
annotation.gr <- makeGRangesFromDataFrame(getAnnotation(GMset), keep.extra.columns = T, start.field = "pos", end.field = "pos")

methdata_to_granges(methdata = Beta, annotation.gr = annotation.gr, chr = "chr6", start_region = 70672841, end_region = 70672878)
}
\author{
Andrew Y.F. Li Yim
}
\keyword{cpg,}
\keyword{dmr,}
\keyword{granges}
\keyword{methylation,}
