% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minfi_to_GEO.R
\name{minfi_to_GEO}
\alias{minfi_to_GEO}
\title{minfi_to_GEO}
\usage{
minfi_to_GEO(values, detP, unmeth, meth)
}
\arguments{
\item{values}{A matrix or dataframe containing the methylation values (i.e. Beta or M-values)}

\item{detP}{A matrix containing the detection P values per probe}

\item{unmeth}{A matrix containing the unmethylated values per probe with the appropriate row names}

\item{meth}{A matrix containing the methylated values per probe with the appropriate column names}
}
\value{
A list of two dataframes: 1) The unmethylated, methylated and detection P-value per sample per probe, and 2) The degree of methylation (Beta, M-value) and the detection P-value per sample per probe.
}
\description{
Provide a GEO submission for Illumina methylation data according to the template provided with GPL16304
}
\examples{

#"I generally provide GEO the processed beta/M-values, and the unprocessed methylated and unmethylated channels as unprocessed data. The detection P-values should be the same regardless of processing"

#Load data
require(minfiData)
baseDir <- system.file("extdata", package = "minfiData")
targets <- read.metharray.sheet(base = baseDir)
RGset <- read.metharray.exp(targets = targets, recursive = T)

#Preprocess it raw to get the raw methylated and unmethylated channels
Mset.raw <- preprocessRaw(RGset)
meth.raw <- getMeth(Mset.raw)
unmeth.raw <- getUnmeth(Mset.raw)
detP.raw <- detectionP(RGset)

#Preprocess it using your normalization tool of interest
Mset <- preprocessIllumina(rgSet = RGset, bg.correct = T, normalize = "controls", reference = 2)
Rset <- ratioConvert(Mset)
GMset <- mapToGenome(Rset)

#Beta values
Beta.processed <- getBeta(GMset)

geo.submission <- minfi_to_GEO(Beta.processed, detP.raw, unmeth.raw, meth.raw)

}
\author{
Andrew Y.F. Li Yim
}
\keyword{450k,}
\keyword{EPIC,}
\keyword{GEO}
\keyword{cpg,}
\keyword{methylation,}
