% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVD_correlator.R
\name{SVD_correlator}
\alias{SVD_correlator}
\title{SVD_correlator}
\usage{
SVD_correlator(SVD, confounder, alpha_threshold = 0.05, padj.method = "BH",
  title)
}
\arguments{
\item{SVD}{Either the output of SVD: A list containing three outputs: left singular matrix u, vector d and right singular matrix v, or a vector containing the right singular matrix v obtained from SVD.}

\item{confounder}{A vector containing the confounder of interest. Order and length must be the same as the number of columns in the right singular matrix v.}

\item{alpha}{A value depicting the alpha threshold used for defining significance. If NULL, then no significance will be provided. Defaults to 0.05.}
}
\value{
A scatterplot and data.frame depicting the correlation of the principal components with the confounders.
}
\description{
Correlates the principal components with a given confounder. In short, the right singular matrix is correlated column per column with the confounder.
}
\author{
Andrew Y.F. Li Yim
}
\keyword{PCA,}
\keyword{SVD,}
\keyword{correlation}
