% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{volcano_plot}
\usage{
volcano_plot(effect_sizes, pvals, identifiers, int_effect_threshold = NULL,
  pval_threshold = 0.05, effect_limit = NULL, top_names = NULL, title)
}
\arguments{
\item{effect_sizes}{A numeric vector of effect sizes}

\item{pvals}{A numeric vector of (adjusted) p values)}

\item{identifiers}{A character vector of identifiers}

\item{int_effect_threshold}{A vertical dashed line to indicate results above an effect size. Defaults to NULL.}

\item{pval_threshold}{A horizontal dashed line to indicate the statistically significant results. Defaults to 0.05}

\item{effect_limit}{If you know there is an upper limit to the effect size. Defaults to finding the limits automatically}

\item{top_names}{A numeric value representing whether you want to plot the names of the top (based on significance) X hits. Defaults to NULL.}

\item{title}{Title of the plot.}
}
\value{
A volcano plot with the effect size on the X axis and the -log10 adjusted p-value on the Y axis.
}
\description{
This function was written since the volcanoplot is an easy way to globally assess the distribution of the p values relative to the effect sizes. This specific function has been tailored to my need of working with methylation data.
}
\examples{
#Load data
require(minfiData)
baseDir <- system.file("extdata", package = "minfiData")
targets <- read.metharray.sheet(base = baseDir)
RGset <- read.metharray.exp(targets = targets, recursive = T)
Mset <- preprocessIllumina(rgSet = RGset, bg.correct = T, normalize = "controls", reference = 2)
Rset <- ratioConvert(Mset)
GMset <- mapToGenome(Rset)

Beta <- getBeta(GMset)
design <- model.matrix(~targets$Sample_Group)

#Linear regression
lfit <- lmFit(Beta, design)
lfit <- eBayes(lfit)
top_genes <- toptable(fit = lfit, coef = 2, number = Inf)

#Plot the volcano_plot

}
\author{
Andrew Y.F. Li Yim
}
\keyword{volcanoplot}

