% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multenrich_plot.R
\name{multenrich_plot}
\alias{multenrich_plot}
\title{multenrich_plot}
\usage{
multenrich_plot(
  enrich_list,
  title = NULL,
  pathway_col,
  es_col,
  pval_col,
  padj_col = NULL,
  alpha = 0.05
)
}
\arguments{
\item{enrich_list}{List of enrichment analyses, wherein each enrichment analysis requires columns containing the pathway name, the enrichment scores (make this 1 or -1 if no particular enrichment score is provided), and a p-value.}

\item{title}{A title for the plot}

\item{pathway_col}{The column name of the pathway columns}

\item{es_col}{The column name of the enrichment scores. Make this 1 or -1 if no particular enrichment score is provided.}

\item{pval_col}{The column name of the p-values columns. One could enter the adjusted p-values here as well, yet I advise not to do so as they tend to flatten out (especially Benjamini-Hochberg adjusted p-values).}

\item{padj_col}{(optional) The column name of the adjusted p-values. This will be used for identifying significant hits, with the non-significant hits being plotted as transparant.}

\item{alpha}{(default: 0.05) The alpha at which you define something as significant.}
}
\value{
A ggplot object containing of heatmap-like plot of the enriched pathways indicating direction, size of the enrichment and the p-value for all provided enrichment analyses.
}
\description{
Produces a boxplot overlayed by a heatmap-like plot for multiple enrichment analyses providing a clear overview of the enrichment scores and p-values when having performed multiple enrichment analyses (i.e. different contrasts).
It is advisable to ensure that the number of pathways to be plotted is less than 100 or else the list will become very long.
The concept and implementation was derived from Martin Garrido Rodriguez-Cordoba's code (https://mgrcbioinfo.github.io/my_GSEA_plot/) and was modified by me.
}
\author{
Martin Garrido Rodriguez-Cordoba, modified by Andrew Y.F. Li Yim
}
\keyword{analysis,}
\keyword{enrichment}
\keyword{heatmap}
\keyword{plot,}
