% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_strip_plot.R
\name{transcript_strip_plot}
\alias{transcript_strip_plot}
\title{transcript_strip_plot}
\usage{
transcript_strip_plot(
  id,
  counts,
  factor_interest,
  title,
  enlarged = "auto",
  type = c("boxplot", "CI", "SE", "mean"),
  colors = NULL,
  legend = TRUE,
  y_lab = NULL
)
}
\arguments{
\item{id}{Identifier as provided in the count matrix}

\item{counts}{Matrix of counts with the identifiers as rownames}

\item{factor_interest}{A vector based on which you want to stratify your plot}

\item{title}{A title for the plot}

\item{enlarged}{Do you want an enlarged plot on the right side (TRUE or FALSE). Defaults to "auto" whereby the standard deviation will be calculated of the sample. If the sd is smaller than 0.015 the enlarged plot will be generated for visibility purposes.}

\item{type}{Choose the type of plot you want: "boxplot", "SE", "CI" or simply "mean". Defaults to "boxplot".}

\item{colors}{A vector containing the the colors to be utilized for the plot with the contrast of interest as names (default: uses ggplot's default coloring scheme).}

\item{legend}{Boolean depicting whether a legend should be included (default: TRUE)}

\item{y_lab}{A specific name for the y labels (default: "Counts")}
}
\value{
A stripplot of a specific gene
}
\description{
Produces a a stripplot of a specific transcript given a count matrix. Ensure that the count matrix has the identifier as the row names
}
\author{
Andrew Y.F. Li Yim
}
\keyword{boxplot,}
\keyword{expression}
\keyword{stripplot,}
