% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{volcano_plot}
\usage{
volcano_plot(effect_sizes, pvals, significance, identifiers,
  int_effect_threshold = NULL, effect_limit = NULL, top_names = NULL,
  title = NULL, x_lab = NULL, y_lab = NULL)
}
\arguments{
\item{effect_sizes}{A numeric vector of effect sizes.}

\item{pvals}{A numeric vector of p values.}

\item{significance}{A boolean vector indicating the significant hits (T = significant, F = non-significant).}

\item{identifiers}{A character vector of identifiers.}

\item{int_effect_threshold}{A vertical dashed line to indicate results above an effect size. Defaults to NULL.}

\item{effect_limit}{Upper limit of the plot. Defaults to finding the limits automatically.}

\item{top_names}{An integer representing the names of the top (based on significance) X hits. Defaults to NULL.}

\item{title}{Title of the plot.}
}
\value{
A volcano plot with the effect size on the X axis and the -log10 p-value on the Y axis. Adjusted p-values are indicated with a different color.
}
\description{
This function was written since the volcanoplot is an easy way to globally assess the distribution of the p values relative to the effect sizes. This specific function has been tailored to my need of working with methylation data.
}
\examples{
#Load data
require(minfiData)
baseDir <- system.file("extdata", package = "minfiData")
targets <- read.metharray.sheet(base = baseDir)
RGset <- read.metharray.exp(targets = targets, recursive = T)
Mset <- preprocessIllumina(rgSet = RGset, bg.correct = T, normalize = "controls", reference = 2)
Rset <- ratioConvert(Mset)
GMset <- mapToGenome(Rset)

Beta <- getBeta(GMset)
design <- model.matrix(~targets$Sample_Group)

#Linear regression
lfit <- lmFit(Beta, design)
lfit <- eBayes(lfit)
top_genes <- toptable(fit = lfit, coef = 2, number = Inf)

#Plot the volcano_plot

}
\author{
Andrew Y.F. Li Yim
}
\seealso{
\pkg{\link{minfi}}

\pkg{\link{minfiData}}
}
\keyword{volcanoplot}

